Improve entry API
