+++ crates/db/src/migration/entry.rs
@@ -1,0 +1,130 @@
+use std::fmt;
+
+use anyhow::{Context, Result, ensure};
+use rusqlite::{Connection, Transaction};
+
+use super::schema::{self, SchemaHash};
+
+/// A migration entry that can be executed inside a SQLite transaction.
+pub(super) trait MigrationEntry {
+    /// Returns the migration name used in diagnostics.
+    fn name(&self) -> &'static str;
+
+    /// Executes the migration body inside `tx`.
+    fn execute_migration(&self, tx: &Transaction<'_>) -> Result<()>;
+}
+
+/// A pure SQL migration.
+pub(super) struct SqlMigration {
+    name: &'static str,
+    sql: &'static str,
+}
+
+impl SqlMigration {
+    pub(super) fn new(name: &'static str, sql: &'static str) -> Self {
+        Self { name, sql }
+    }
+}
+
+impl MigrationEntry for SqlMigration {
+    fn name(&self) -> &'static str {
+        self.name
+    }
+
+    fn execute_migration(&self, tx: &Transaction<'_>) -> Result<()> {
+        tx.execute_batch(self.sql).map_err(Into::into)
+    }
+}
+
+impl fmt::Debug for SqlMigration {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("SqlMigration").field("name", &self.name).finish_non_exhaustive()
+    }
+}
+
+/// A Rust migration function.
+pub(super) struct CodeMigration {
+    name: &'static str,
+    apply: CodeMigrationFn,
+}
+
+impl CodeMigration {
+    pub(super) fn new(name: &'static str, apply: CodeMigrationFn) -> Self {
+        Self { name, apply }
+    }
+}
+
+impl MigrationEntry for CodeMigration {
+    fn name(&self) -> &'static str {
+        self.name
+    }
+
+    fn execute_migration(&self, tx: &Transaction<'_>) -> Result<()> {
+        (self.apply)(tx)
+    }
+}
+
+impl fmt::Debug for CodeMigration {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("CodeMigration")
+            .field("name", &self.name)
+            .finish_non_exhaustive()
+    }
+}
+
+/// Applies `migration`, sets `user_version`, commits, and returns the resulting schema hash.
+pub(super) fn apply_migration(
+    conn: &mut Connection,
+    version: usize,
+    migration: &impl MigrationEntry,
+) -> Result<SchemaHash> {
+    apply_migration_transaction(conn, version, migration, Ok::<SchemaHash, anyhow::Error>)
+}
+
+/// Applies `migration`, verifies the resulting schema hash, sets `user_version`, and commits.
+pub(super) fn apply_migration_and_verify_schema(
+    conn: &mut Connection,
+    version: usize,
+    migration: &impl MigrationEntry,
+    expected: SchemaHash,
+) -> Result<()> {
+    let name = migration.name();
+    apply_migration_transaction(conn, version, migration, |actual| {
+        ensure!(
+            actual == expected,
+            "schema hash mismatch after migration {version} \"{name}\": expected {expected}, got \
+             {actual}"
+        );
+        Ok(())
+    })
+}
+
+fn apply_migration_transaction<T>(
+    conn: &mut Connection,
+    version: usize,
+    migration: &impl MigrationEntry,
+    verify_hash: impl FnOnce(SchemaHash) -> Result<T>,
+) -> Result<T> {
+    let name = migration.name();
+    let tx = conn.transaction().with_context(|| {
+        format!("failed to start transaction for migration {version} \"{name}\"")
+    })?;
+
+    migration
+        .execute_migration(&tx)
+        .with_context(|| format!("failed to apply migration {version} \"{name}\""))?;
+    let schema_hash = SchemaHash::new(&tx).with_context(|| {
+        format!("failed to compute schema hash after migration {version} \"{name}\"")
+    })?;
+    let result = verify_hash(schema_hash)?;
+    schema::set_version(&tx, version).with_context(|| {
+        format!("failed to update user_version for migration {version} \"{name}\"")
+    })?;
+    tx.commit()
+        .with_context(|| format!("failed to commit migration {version} \"{name}\""))?;
+
+    Ok(result)
+}
+
+/// A Rust migration function executed inside a SQLite transaction.
+pub type CodeMigrationFn = for<'conn> fn(&Transaction<'conn>) -> Result<()>;
